services/matching-engine/requirements.txt
faust-streaming==0.10.14
aiohttp==3.9.1
prometheus-client==0.19.0
python-dotenv==1.0.0
structlog==23.3.0
tenacity==8.2.3
